refactor: moved new source to extension folder
