+++ src/extension/sources/ColorFormat.ts
